complete: 完善功能，达到npm包级别并进行发布
